+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - none
 
 # v2023.1.1-beta
+- ci - auto-create asset_image_logo_512.png from asset_image_logo_512.html
 - replace branch shell with branch sh_lin, sh_mac, sh_win
 - cron - add loopback to mycron2 stub
 - jslint-ci - revamp auto-updating and add shell-function shGithubCheckoutRemote